pyproject.toml
[tool.poetry]
name = "newproj"
version = "0.1.0"
description = ""
authors = ["Sudharsan Rangarajan <sudharsan.rangarajan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
llama-index = "^0.6.6"
python-dotenv = "^1.0.0"
pypdf2 = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
